[SC script]: Convert the uint256[] to the uint8[] in the for-loop of the PubkeyAndSignedMessageExtractor.sol

+++ scripts/utils/array-bytes-generator/contracts/libraries/PubkeyAndSignedMessageExtractor.sol
@@ -51,32 +51,36 @@ contract PubkeyAndSignedMessageExtractor is Script {
         console.log("insurer_pubkey_bytes_array.length: %s", insurer_pubkey_bytes_array.length); // [Log]: 64
 
         /// @dev - Store a Uint8Array value, which was retrieved from the output.json, into the uint256 array variable (uint256[])
-        uint8[] memory _insurer_signature_bytes = vm.parseJsonUintArray(json, ".insurer_signature_bytes");
-        //uint256[] memory _insurer_signature_bytes = vm.parseJsonUintArray(json, ".insurer_signature_bytes");
+        //uint8[] memory _insurer_signature_bytes = vm.parseJsonUintArray(json, ".insurer_signature_bytes");
+        uint256[] memory _insurer_signature_bytes = vm.parseJsonUintArray(json, ".insurer_signature_bytes");
         console.log("_insurer_signature_bytes.length: %s", _insurer_signature_bytes.length);     // [Log]: 99
         for (uint i = 0; i < _insurer_signature_bytes.length - 1; i++) {
-            insurer_signature_bytes_array[i] = _insurer_signature_bytes[i];
+            insurer_signature_bytes_array[i] = uint8(_insurer_signature_bytes[i]);
+            //insurer_signature_bytes_array[i] = _insurer_signature_bytes[i];
             console.log("insurer_signature_bytes_array[%s] = %s", i, insurer_signature_bytes_array[i]); // [Log - Success]: _insurer_signature_bytes[0] = 211, _insurer_signature_bytes[1] = 23, ...
         }
 
-        uint8[] memory _hospital_signature_bytes = vm.parseJsonUintArray(json, ".hospital_signature_bytes");
-        //uint256[] memory _hospital_signature_bytes = vm.parseJsonUintArray(json, ".hospital_signature_bytes");
+        //uint8[] memory _hospital_signature_bytes = vm.parseJsonUintArray(json, ".hospital_signature_bytes");
+        uint256[] memory _hospital_signature_bytes = vm.parseJsonUintArray(json, ".hospital_signature_bytes");
         for (uint i = 0; i < _hospital_signature_bytes.length - 1; i++) {
-            hospital_signature_bytes_array[i] = _hospital_signature_bytes[i];
+            hospital_signature_bytes_array[i] = uint8(_hospital_signature_bytes[i]);
+            //hospital_signature_bytes_array[i] = _hospital_signature_bytes[i];
             console.log("hospital_signature_bytes_array[%s] = %s", i, hospital_signature_bytes_array[i]); // [Log - Success]: _hospital_signature_bytes[0] = 211, _hospital_signature_bytes[1] = 23, ...
         }
 
-        uint8[] memory _insurer_pubkey_bytes = vm.parseJsonUintArray(json, ".insurer_pubkey_bytes");
-        //uint256[] memory _insurer_pubkey_bytes = vm.parseJsonUintArray(json, ".insurer_pubkey_bytes");
+        //uint8[] memory _insurer_pubkey_bytes = vm.parseJsonUintArray(json, ".insurer_pubkey_bytes");
+        uint256[] memory _insurer_pubkey_bytes = vm.parseJsonUintArray(json, ".insurer_pubkey_bytes");
         for (uint i = 0; i < _insurer_pubkey_bytes.length; i++) {
-            insurer_pubkey_bytes_array[i] = _insurer_pubkey_bytes[i];
+            insurer_pubkey_bytes_array[i] = uint8(_insurer_pubkey_bytes[i]);
+            //insurer_pubkey_bytes_array[i] = _insurer_pubkey_bytes[i];
             console.log("insurer_pubkey_bytes_array[%s] = %s", i, insurer_pubkey_bytes_array[i]); // [Log - Success]: _insurer_pubkey_bytes[0] = 0x215597bacd9c7e977dfc170f320074155de974be494579d2586e5b268fa3b629, _insurer_pubkey_bytes[1] = 0x26df0d347e961cb94e1cc6d2ad8558696de8c1964b30e26f2ec8b926cbbbf862, ...
         }
 
-        uint8[] memory _hospital_pubkey_bytes = vm.parseJsonUintArray(json, ".hospital_pubkey_bytes");
-        //uint256[] memory _hospital_pubkey_bytes = vm.parseJsonUintArray(json, ".hospital_pubkey_bytes");
+        //uint8[] memory _hospital_pubkey_bytes = vm.parseJsonUintArray(json, ".hospital_pubkey_bytes");
+        uint256[] memory _hospital_pubkey_bytes = vm.parseJsonUintArray(json, ".hospital_pubkey_bytes");
         for (uint i = 0; i < _hospital_pubkey_bytes.length - 1; i++) {
-            hospital_pubkey_bytes_array[i] = _hospital_pubkey_bytes[i];
+            hospital_pubkey_bytes_array[i] = uint8(_hospital_pubkey_bytes[i]);
+            //hospital_pubkey_bytes_array[i] = _hospital_pubkey_bytes[i];
             console.log("hospital_signature_bytes_array[%s] = %s", i, hospital_signature_bytes_array[i]); // [Log - Success]: _hospital_pubkey_bytes[0] = 0x215597bacd9c7e977dfc170f320074155de974be494579d2586e5b268fa3b629, _hospital_pubkey_bytes[1] = 0x26df0d347e961cb94e1cc6d2ad8558696de8c1964b30e26f2ec8b926cbbbf862, ...
         }
